requirements.txt
django
djangorestframework
django-cors-headers
django-filter
djangorestframework-simplejwt
psycopg2-binary
python-decouple
pillow
django-ckeditor-5
gunicorn
requests
termcolor
